backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==3.2.2
cffi==2.0.0
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
ecdsa==0.19.1
fastapi==0.128.0
fonttools==4.61.1
graphviz==0.21
h11==0.16.0
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
narwhals==2.15.0
numpy==2.4.1
packaging==26.0
pandas==3.0.0
passlib==1.7.4
pillow==12.1.0
plotly==6.5.2
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-multipart==0.0.22
rsa==4.9.1
scipy==1.17.0
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
google-generativeai==0.8.3
python-dotenv==1.0.1
websockets==14.1